+++ export_academy/views.py
@@ -10,7 +10,7 @@ from django_filters.views import FilterView
 
 from config import settings
 from core import mixins as core_mixins
-from export_academy import filters, forms, models
+from export_academy import filters, forms, helpers, models
 from export_academy.helpers import get_buttons_for_event
 from export_academy.mixins import BookingMixin
 from export_academy.models import ExportAcademyHomePage
@@ -31,19 +31,13 @@ class EventListView(
         user = self.request.user
         return get_buttons_for_event(user, event)
 
+    def get(self, request, *args, **kwargs):
+        request.GET = helpers.build_request_navigation_params(request)
+        return super().get(request, *args, **kwargs)
+
     def get_context_data(self, **kwargs):
         ctx = super().get_context_data(**kwargs)
-        user = self.request.user
-        bookings = []
-
-        if user.is_authenticated:
-            bookings = models.Booking.objects.filter(
-                registration_id=user.email, status='Confirmed'  # type: ignore
-            ).values_list('event_id', flat=True)
-
-        ctx.update(bookings=bookings, filter=self.filterset_class(self.request.GET))
         ctx['landing_page'] = ExportAcademyHomePage.objects.first()
-
         return ctx